+++ include/game.h
@@ -5,7 +5,7 @@
 #include <glad/glad.h>
 #include <GLFW/glfw3.h>
 #include <iostream>
-#include "../piece/piece.h"
+#include "piece.h"
 #include <memory>
 
 class Render;
