+++ src/shader.cpp
@@ -1,4 +1,4 @@
-#include "shader.h"
+#include "../include/shader.h"
 #include <glm/ext/matrix_float4x4.hpp>
 #include <glm/gtc/type_ptr.hpp>
